Fixes for 3D rendering + Take model scaling and offsets into account when rendering (GUI) * Fixed a bug where using UI scaling would not scale the GlControl viewport * Reduced crashes when rendering 3D models in Windows (temporary fix, opening the about dialog still crashes the program as mentioned in #1)

+++ FlipnicFileToolGUI/Shaders/UiOpenGlShader.cs
@@ -84,6 +84,10 @@ namespace FlipnicFileToolGUI.Shaders
 
         ~UiOpenGlShader()
         {
+            if (OperatingSystem.IsWindows())
+            {
+                return;
+            }
             GL.DeleteProgram(_handle);
         }
